[infra] rename server -> game, add random agent
