+++ pkg/infra/game/state/state.go
